Change Timer screen layout
