+++ app/src/main/java/com/bnyro/clock/ui/screens/TimerScreen.kt
@@ -1,38 +1,31 @@
 package com.bnyro.clock.ui.screens
 
+import androidx.compose.foundation.ExperimentalFoundationApi
+import androidx.compose.foundation.background
+import androidx.compose.foundation.combinedClickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.LazyRow
-import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.lazy.grid.GridCells
+import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
+import androidx.compose.foundation.lazy.grid.items
 import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowBack
-import androidx.compose.material.icons.filled.Close
 import androidx.compose.material.icons.filled.Create
-import androidx.compose.material.icons.filled.Edit
-import androidx.compose.material.icons.filled.Notifications
-import androidx.compose.material.icons.filled.Pause
 import androidx.compose.material.icons.filled.PlayArrow
-import androidx.compose.material3.AlertDialog
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.ElevatedCard
 import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.Icon
-import androidx.compose.material3.LinearProgressIndicator
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SmallFloatingActionButton
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -43,25 +36,19 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.StrokeCap
+import androidx.compose.ui.draw.clip
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
-import com.bnyro.clock.R
-import com.bnyro.clock.extensions.addZero
 import com.bnyro.clock.obj.NumberKeypadOperation
-import com.bnyro.clock.obj.WatchState
 import com.bnyro.clock.ui.common.ExampleTimer
-import com.bnyro.clock.ui.components.ClickableIcon
-import com.bnyro.clock.ui.components.DialogButton
 import com.bnyro.clock.ui.components.FormattedTimerTime
 import com.bnyro.clock.ui.components.NumberKeypad
-import com.bnyro.clock.ui.dialog.RingtonePickerDialog
-import com.bnyro.clock.ui.components.ScrollTimePicker
+import com.bnyro.clock.ui.components.TimePickerDial
+import com.bnyro.clock.ui.components.TimerItem
 import com.bnyro.clock.ui.model.TimerModel
 import com.bnyro.clock.util.Preferences
 
+@OptIn(ExperimentalFoundationApi::class)
 @Composable
 fun TimerScreen(timerModel: TimerModel) {
     val context = LocalContext.current
@@ -75,80 +62,63 @@ fun TimerScreen(timerModel: TimerModel) {
         mutableStateOf(false)
     }
 
-    Box(
-        modifier = Modifier.fillMaxSize(),
-        contentAlignment = Alignment.Center
-    ) {
+    Scaffold(floatingActionButton = {
         if (timerModel.scheduledObjects.isEmpty() || createNew) {
             Column(
+                horizontalAlignment = Alignment.CenterHorizontally,
+                verticalArrangement = Arrangement.spacedBy(8.dp)
+            ) {
+                if (timerModel.scheduledObjects.isNotEmpty()) {
+                    SmallFloatingActionButton(
+                        containerColor = MaterialTheme.colorScheme.tertiaryContainer,
+                        onClick = { createNew = false }
+                    ) {
+                        Icon(imageVector = Icons.Default.ArrowBack, contentDescription = null)
+                    }
+                }
+                FloatingActionButton(
+                    onClick = {
+                        createNew = false
+                        timerModel.startTimer(context)
+                    }
+                ) {
+                    Icon(imageVector = Icons.Default.PlayArrow, contentDescription = null)
+                }
+            }
+        } else {
+            FloatingActionButton(
+                modifier = Modifier,
+                onClick = { createNew = true }
+            ) {
+                Icon(imageVector = Icons.Default.Create, contentDescription = null)
+            }
+        }
+    }) { paddingValues ->
+        if (timerModel.scheduledObjects.isEmpty() || createNew) {
+            Column(
+                modifier = Modifier
+                    .fillMaxSize()
+                    .padding(paddingValues),
                 horizontalAlignment = Alignment.CenterHorizontally
             ) {
                 if (useOldPicker) {
-                    Column(
-                        modifier = Modifier.weight(1f),
-                        verticalArrangement = Arrangement.Center,
-                        horizontalAlignment = Alignment.CenterHorizontally
+                    Row(
+                        Modifier
+                            .fillMaxWidth()
+                            .weight(2f),
+                        horizontalArrangement = Arrangement.Center,
+                        verticalAlignment = Alignment.CenterVertically
                     ) {
-                        Row(
-                            verticalAlignment = Alignment.CenterVertically,
-                            horizontalArrangement = Arrangement.spacedBy(40.dp)
-                        ) {
-                            Text(
-                                text = stringResource(R.string.hours),
-                                style = MaterialTheme.typography.bodyMedium,
-                                color = MaterialTheme.colorScheme.primary
-                            )
-                            Text(
-                                text = stringResource(R.string.minutes),
-                                style = MaterialTheme.typography.bodyMedium,
-                                color = MaterialTheme.colorScheme.primary
-                            )
-                            Text(
-                                text = stringResource(R.string.seconds),
-                                style = MaterialTheme.typography.bodyMedium,
-                                color = MaterialTheme.colorScheme.primary
-                            )
-                        }
-                        Spacer(modifier = Modifier.height(32.dp))
-                        Row(
-                            verticalAlignment = Alignment.CenterVertically,
-                            horizontalArrangement = Arrangement.spacedBy(16.dp)
-                        ) {
-                            ScrollTimePicker(
-                                value = timerModel.getHours(),
-                                onValueChanged = timerModel::addHours,
-                                maxValue = 24
-                            )
-                            Text(
-                                text = ":",
-                                style = MaterialTheme.typography.displayMedium,
-                                color = MaterialTheme.colorScheme.primary
-                            )
-                            ScrollTimePicker(
-                                value = timerModel.getMinutes(),
-                                onValueChanged = timerModel::addMinutes,
-                                maxValue = 60
-                            )
-                            Text(
-                                text = ":",
-                                style = MaterialTheme.typography.displayMedium,
-                                color = MaterialTheme.colorScheme.primary
-                            )
-                            ScrollTimePicker(
-                                value = timerModel.getSeconds(),
-                                onValueChanged = timerModel::addSeconds,
-                                maxValue = 60
-                            )
-                        }
+                        TimePickerDial(timerModel)
                     }
                 } else {
                     Column(
-                        modifier = Modifier.weight(1f),
+                        modifier = Modifier.weight(3f),
                         horizontalAlignment = Alignment.CenterHorizontally,
                         verticalArrangement = Arrangement.Center
                     ) {
                         FormattedTimerTime(
-                            seconds = timerModel.timePickerSecondsState.toInt(),
+                            seconds = timerModel.timePickerFakeUnits,
                             modifier = Modifier.padding(bottom = 32.dp)
                         )
                         NumberKeypad(
@@ -165,194 +135,52 @@ fun TimerScreen(timerModel: TimerModel) {
                         )
                     }
                 }
-                Box(modifier = Modifier.align(Alignment.End)) {
-                    if (timerModel.scheduledObjects.isNotEmpty()) {
-                        SmallFloatingActionButton(
-                            modifier = Modifier.padding(end = 16.dp),
-                            containerColor = MaterialTheme.colorScheme.tertiaryContainer,
-                            onClick = { createNew = false }
-                        ) {
-                            Icon(imageVector = Icons.Default.ArrowBack, contentDescription = null)
-                        }
-                    }
-                }
-                Row(
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Column(modifier = Modifier.weight(1f)) {
-                        LazyRow(
-                            contentPadding = PaddingValues(start = 40.dp),
-                            horizontalArrangement = Arrangement.spacedBy(16.dp)
-                        ) {
-                            if (showExampleTimers) {
-                                items(ExampleTimer.exampleTimers) { timer ->
-                                    Button(
+                if (showExampleTimers) {
+                    LazyVerticalGrid(
+                        modifier = Modifier
+                            .fillMaxWidth()
+                            .weight(1f),
+                        columns = GridCells.Adaptive(100.dp),
+                        contentPadding = PaddingValues(16.dp),
+                        verticalArrangement = Arrangement.spacedBy(8.dp),
+                        horizontalArrangement = Arrangement.spacedBy(8.dp)
+                    ) {
+                        items(items = ExampleTimer.exampleTimers) { timer ->
+                            Box(
+                                modifier = Modifier
+                                    .clip(RoundedCornerShape(16.dp))
+                                    .combinedClickable(
                                         onClick = {
-                                            timerModel.setSeconds(timer.seconds)
+                                            timerModel.timePickerSeconds = timer.seconds
                                             createNew = false
                                             timerModel.startTimer(context)
-                                        },
-                                        colors = ButtonDefaults.buttonColors()
-                                    ) {
-                                        Text(
-                                            timer.formattedTime
-                                        )
-                                    }
-                                }
+                                        }
+                                    )
+                                    .width(100.dp)
+                                    .background(MaterialTheme.colorScheme.secondaryContainer)
+                                    .padding(8.dp),
+                                contentAlignment = Alignment.Center
+                            ) {
+                                Text(
+                                    timer.formattedTime,
+                                    style = MaterialTheme.typography.titleLarge,
+                                    color = MaterialTheme.colorScheme.onSecondaryContainer
+                                )
                             }
                         }
-                    }
-                    FloatingActionButton(
-                        modifier = Modifier.padding(16.dp),
-                        onClick = {
-                            createNew = false
-                            timerModel.startTimer(context)
-                        }
-                    ) {
-                        Icon(imageVector = Icons.Default.PlayArrow, contentDescription = null)
                     }
                 }
             }
         } else {
             LazyColumn(
-                modifier = Modifier.fillMaxSize(),
+                modifier = Modifier
+                    .fillMaxSize()
+                    .padding(paddingValues),
                 verticalArrangement = Arrangement.Top
             ) {
                 itemsIndexed(timerModel.scheduledObjects) { index, obj ->
-                    val minutes = obj.currentPosition.value / 60000
-                    val seconds = (obj.currentPosition.value % 60000) / 1000
-                    val hundreds = obj.currentPosition.value % 1000 / 10
-
-                    var showLabelEditor by remember {
-                        mutableStateOf(false)
-                    }
-                    var showRingtoneEditor by remember {
-                        mutableStateOf(false)
-                    }
-
-                    ElevatedCard(
-                        modifier = Modifier.padding(horizontal = 12.dp, vertical = 6.dp),
-                        shape = RoundedCornerShape(20.dp)
-                    ) {
-                        Column {
-                            Row(
-                                modifier = Modifier.padding(
-                                    horizontal = 16.dp,
-                                    vertical = 16.dp
-                                )
-                            ) {
-                                Column {
-                                    obj.label.value?.let { label ->
-                                        Text(
-                                            text = label,
-                                            style = MaterialTheme.typography.headlineLarge,
-                                            maxLines = 1,
-                                            overflow = TextOverflow.Ellipsis
-                                        )
-                                    }
-                                    Row(
-                                        verticalAlignment = Alignment.Bottom
-                                    ) {
-                                        Text(
-                                            text = "$minutes:${seconds.addZero()}",
-                                            style = MaterialTheme.typography.headlineMedium
-                                        )
-                                        Spacer(modifier = Modifier.width(5.dp))
-                                        Text(
-                                            text = hundreds.addZero()
-                                        )
-                                    }
-                                }
-                                Spacer(modifier = Modifier.weight(1f))
-                                ClickableIcon(imageVector = Icons.Default.Edit) {
-                                    showLabelEditor = true
-                                }
-                                ClickableIcon(imageVector = Icons.Default.Notifications) {
-                                    showRingtoneEditor = true
-                                }
-                                ClickableIcon(imageVector = Icons.Default.Close) {
-                                    timerModel.stopTimer(context, index)
-                                }
-                                FloatingActionButton(onClick = {
-                                    when (obj.state.value) {
-                                        WatchState.PAUSED -> timerModel.resumeTimer(index)
-                                        WatchState.RUNNING -> timerModel.pauseTimer(index)
-                                        else -> timerModel.startTimer(context)
-                                    }
-                                }) {
-                                    Icon(
-                                        imageVector = if (obj.state.value == WatchState.RUNNING) {
-                                            Icons.Default.Pause
-                                        } else {
-                                            Icons.Default.PlayArrow
-                                        },
-                                        contentDescription = null
-                                    )
-                                }
-                            }
-                            LinearProgressIndicator(
-                                modifier = Modifier.fillMaxWidth().padding(8.dp).height(10.dp),
-                                progress = obj.currentPosition.value / obj.initialPosition.toFloat(),
-                                strokeCap = StrokeCap.Round
-                            )
-                        }
-                    }
-
-                    if (showLabelEditor) {
-                        var newLabel by remember {
-                            mutableStateOf(obj.label.value.orEmpty())
-                        }
-
-                        AlertDialog(
-                            onDismissRequest = { showLabelEditor = false },
-                            confirmButton = {
-                                DialogButton(android.R.string.ok) {
-                                    timerModel.service?.updateLabel(obj.id, newLabel)
-                                    newLabel = ""
-                                    showLabelEditor = false
-                                }
-                            },
-                            dismissButton = {
-                                DialogButton(android.R.string.cancel) {
-                                    newLabel = ""
-                                    showLabelEditor = false
-                                }
-                            },
-                            title = {
-                                Text(stringResource(R.string.label))
-                            },
-                            text = {
-                                OutlinedTextField(
-                                    value = newLabel,
-                                    onValueChange = { newLabel = it },
-                                    label = {
-                                        Text(stringResource(R.string.label))
-                                    }
-                                )
-                            }
-                        )
-                    }
-
-                    if (showRingtoneEditor) {
-                        RingtonePickerDialog(onDismissRequest = {
-                            showRingtoneEditor = false
-                        }) { _, uri ->
-                            timerModel.service?.updateRingtone(obj.id, uri)
-                        }
-                    }
+                    TimerItem(obj, index, timerModel)
                 }
-
-                item {
-                    Spacer(modifier = Modifier.height(100.dp))
-                }
-            }
-            FloatingActionButton(
-                modifier = Modifier
-                    .align(Alignment.BottomEnd)
-                    .padding(16.dp),
-                onClick = { createNew = true }
-            ) {
-                Icon(imageVector = Icons.Default.Create, contentDescription = null)
             }
         }
     }
